requirements.txt
numpy==1.26.4
scipy==1.13.0
matplotlib==3.8.4
flask==2.3.2
jpype1==1.5.0
scikit-learn~=1.7.1
networkx~=3.5